Experiment: inject context at runtime.
By injecting the context when we create the experiment, it meant that
the experiment was always in a "fresh" state. This means that the
counter would always have been reset for the percentage option.

We now inject the context - and this also adds the posibility to add
`nil` - at runtime, which makes it possible to setup the experiment
outside of the concurrency scope but still use it concurrently.