Experiment: major refactor.
This is a major refactor to the structure of how this package works.
This is mainly done to achieve concurrent access, which is obviously
important and was buggy before.

We now create a runner per `Run()` instance, which makes concurrency
easier as it eliminates the state between calls.

This also removes the variadic approach of initialising the experiment
and uses a proper config struct. There is a `DefaultConfig()` function
available.

There are some breaking changes to the API:

- There is no `New` function anymore. You can create a new experiment by
  initialising a new struct, `&Experiment{DefaultConfig("test")}`.
- There is no `Init` function anymore. Everything is done through a
  `Config` struct.
- `Run()` now returns a - newly created - `Observations` type.
- `Experiment` doesn't have a `Result()` method anymore, this will now
  be created by calling `NewResult(Observations, Config)`. (to be
  implemented)
- `Experiment` doesn't have a `Publish()` method anymore, this will be
  found on the `Result` type. (to be implemented)
- The `Observation` interface is gone and has become just a regular
  struct with exported variables.
- `TestMode` is now a package variable instead of a config param.