Context: remove external dependency.
This removes the external dependency of the `/x/net/context` package.
Due to backwards compability reasons, the interface has been copied to
the package. We want to support Go versions below 1.7.

The package now also expects the context to be actually set and won't
create a default if this is not the case. The Go Context package that
one should use a `context.TODO()` when it is unclear which context to
use. The end user should now properly implement this behaviour.