Streamline error handling.
Before, Control and Candidate were differentiated by their return value.
Candidate functions could return an error only if the name was
'control'. This now panics for that, as it is a developer error which we
want to highlight as soon as possible and is not recoverable.

This also introduces struct errors so it can leverage errors.Is and
errors.As properly.